Add Antiwork homepage; move bounties board to /bounties
New marketing homepage at / — "We automated ourselves out of the work."
Frames the model: the agent software is free/open-source, the $10K call is
the only paid offer. Pulls real production proof (support auto-resolution,
fraud/risk autonomy, cycle time, revenue/employee) and links to the live
Gumroad call product.

The existing GitHub bounties board moves from / to /bounties, unchanged.
All "back home" links across the 2024/2025/2026 decks now land on the new
homepage, which is the intended behavior.
